UserInterface/requirements.txt
Flask==2.0.2
google-cloud-storage==2.6.0
pyodbc==4.0.30
Werkzeug==2.0.3
python-dotenv==1.0.0
flask-cors==3.0.10
